swap-3329: added authentication and encryption, added tests if allowed to save message and hdf structure as text

+++ ingestor_lib.py
@@ -245,25 +245,33 @@ def ingest_message(
     path_name = os.path.dirname(full_file_name)
     logger.info('Dataset folder : {}'.format(path_name))
     logger.info('Dataset raw data file : {}'.format(file_name))
+    # create path for files saved by the ingestor
+    ingestor_files_path = os.path.join(os.path.dirname(path_name),config["run_options"]["ingestor_files_folder"])
+    logger.info("Ingestor files folder: {}".format(ingestor_files_path))
 
     # list of files to be added to the dataset
     files_list = []
     if config["run_options"]["message_to_file"]:
-        message_file_name = os.path.splitext(file_name)[0] + config["run_options"]["message_file_extension"]
-        logger.info("message file name : " + message_file_name)
-        message_full_file_path = os.path.join(
-            path_name if config["run_options"]["message_output"] == "SOURCE_FOLDER" else config["run_options"]["local_output_folder"],
-            message_file_name
-        )
-        logger.info("message full file path : " + message_full_file_path)
-        with open(message_full_file_path, 'w') as fh:
-            json.dump(entry, fh)
-        logger.info("message saved to file")
-        if config["run_options"]["message_output"] == "SOURCE_FOLDER":
-            files_list += [{
-                "path": message_full_file_path,
-                "size": len(json.dumps(entry)),
-            }]
+        message_file_path = ingestor_files_path if config["run_options"]["message_output"] == "SOURCE_FOLDER" else config["run_options"]["local_output_folder"]
+        logger.info("message file will be saved in {}".format(message_file_path))
+        if os.path.exists(message_file_path):
+            message_file_name = os.path.splitext(file_name)[0] + config["run_options"]["message_file_extension"]
+            logger.info("message file name : " + message_file_name)
+            message_full_file_path = os.path.join(
+                message_file_path,
+                message_file_name
+            )
+            logger.info("message full file path : " + message_full_file_path)
+            with open(message_full_file_path, 'w') as fh:
+                json.dump(entry, fh)
+            logger.info("message saved to file")
+            if config["run_options"]["message_output"] == "SOURCE_FOLDER":
+                files_list += [{
+                    "path": message_full_file_path,
+                    "size": len(json.dumps(entry)),
+                }]
+        else:
+            logger.info("Message file path not accessible")
 
     if entry.metadata is not None:
         metadata = json.loads(entry.metadata)
@@ -299,20 +307,25 @@ def ingest_message(
         else:
             logger.debug("hdf structure dict : " + json.dumps(hdf_structure_dict))
         if config["run_options"]["hdf_structure_to_file"]:
-            hdf_structure_file_name = os.path.join(
-                path_name \
-                    if config["run_options"]["hdf_structure_output"] == "SOURCE_FOLDER" \
-                    else os.path.abspath(config["run_options"]["files_output_folder"]),
-                os.path.splitext(file_name)[0] + config["run_options"]["hdf_structure_file_extension"])
-            logger.info("hdf structure file name : " + hdf_structure_file_name)
-            with open(hdf_structure_file_name,'w') as fh:
-                json.dump(hdf_structure_dict,fh)
-            logger.info("hdf structure saved to file : " + hdf_structure_file_name)
-            if config["run_options"]["hdf_structure_output"] == "SOURCE_FOLDER":
-                files_list += [{
-                    "path": hdf_structure_file_name,
-                    "size": len(json.dumps(hdf_structure_dict)),
-                }]
+            hdf_structure_file_path = ingestor_files_path \
+                if config["run_options"]["hdf_structure_output"] == "SOURCE_FOLDER" \
+                else os.path.abspath(config["run_options"]["files_output_folder"])
+            logger.info("hdf structure file will be saved in {}".format(hdf_structure_file_path))
+            if os.path.exists(hdf_structure_file_path):
+                hdf_structure_file_name = os.path.join(
+                    hdf_structure_file_path,
+                    os.path.splitext(file_name)[0] + config["run_options"]["hdf_structure_file_extension"])
+                logger.info("hdf structure file name : " + hdf_structure_file_name)
+                with open(hdf_structure_file_name,'w') as fh:
+                    json.dump(hdf_structure_dict,fh)
+                logger.info("hdf structure saved to file : " + hdf_structure_file_name)
+                if config["run_options"]["hdf_structure_output"] == "SOURCE_FOLDER":
+                    files_list += [{
+                        "path": hdf_structure_file_name,
+                        "size": len(json.dumps(hdf_structure_dict)),
+                    }]
+            else:
+                logger.info("hdf structure file path not accessible")
 
         # retrieve proposal id, if present
         proposal_id = None
@@ -638,7 +651,7 @@ def update_file_info(
     output_item = {
         "path": os.path.basename(file_item["path"]),
     }
-    if compute_file_stats:
+    if compute_file_stats and os.path.exists(file_item["path"]):
         stats = os.stat(file_item["path"])
         output_item = {
             **output_item,
@@ -659,7 +672,7 @@ def update_file_info(
             }
         }
 
-    if compute_file_hash:
+    if compute_file_hash and os.path.exists(file_item["path"]):
         output_item["chk"] = checksum_of_file(file_item["path"], file_hash_algorithm)
 
     return output_item
